Modify: unnecessary (uint8_t) casting

+++ src/headers/tlshdr.cpp
@@ -3,7 +3,7 @@
 bool TlsHdr::parse(const uint8_t *pkt, uint32_t len) {
     uint32_t offset = 0;
 
-    contentType_ = (uint8_t)pkt[offset];
+    contentType_ = pkt[offset];
     offset += 3;
     if (contentType() != TlsHdr::handshake)
         return false;
@@ -11,7 +11,7 @@ bool TlsHdr::parse(const uint8_t *pkt, uint32_t len) {
     length_ = (uint16_t)(pkt[offset] << 8 | pkt[offset + 1]);
     offset += 2;
 
-    type_ = (uint8_t)pkt[offset];
+    type_ = pkt[offset];
     offset += 1;
     if (type() != TlsHdr::clientHello)
         return false;
@@ -19,13 +19,13 @@ bool TlsHdr::parse(const uint8_t *pkt, uint32_t len) {
     handshakeLength_ = (uint32_t)(pkt[offset + 1] << 16 | pkt[offset + 2] << 8 | pkt[offset + 3]);
     offset += 37;
 
-    sessionIdLength_ = (uint8_t)pkt[offset];
+    sessionIdLength_ = pkt[offset];
     offset += sessionIdLength() + 1;
 
     cipherSuitesLength_ = (uint16_t)(pkt[offset] << 8 | pkt[offset + 1]);
     offset += cipherSuitesLength() + 2;
 
-    compMethodsLength_ = (uint8_t)pkt[offset];
+    compMethodsLength_ = pkt[offset];
     offset += compMethodsLength() + 1;
 
     extensionsLength_ = (uint16_t)(pkt[offset] << 8 | pkt[offset + 1]);
